+++ jsolex/src/main/java/me/champeau/a4j/jsolex/app/listeners/BatchModeEventListener.java
@@ -378,7 +378,7 @@ public class BatchModeEventListener implements ProcessingEventListener, ImageMat
                 tabPane.getSelectionModel().select(imagesViewerTab);
         });
         result.imagesByLabel().entrySet().stream().parallel().forEach(entry -> {
-            var name = namingStrategy.render(0, null, Constants.TYPE_PROCESSED, entry.getKey(), "batch");
+            var name = namingStrategy.render(0, null, Constants.TYPE_PROCESSED, entry.getKey(), "batch", entry.getValue());
             var outputFile = new File(outputDirectory, name);
             delegate.onImageGenerated(new ImageGeneratedEvent(
                 new GeneratedImage(GeneratedImageKind.IMAGE_MATH, entry.getKey(), outputFile.toPath(), entry.getValue(), null)
@@ -386,7 +386,7 @@ public class BatchModeEventListener implements ProcessingEventListener, ImageMat
             hasCustomImages.set(true);
         });
         result.filesByLabel().entrySet().stream().parallel().forEach(entry -> {
-            var name = namingStrategy.render(0, null, Constants.TYPE_PROCESSED, entry.getKey(), "batch");
+            var name = namingStrategy.render(0, null, Constants.TYPE_PROCESSED, entry.getKey(), "batch", null);
             try {
                 var fileName = entry.getValue().toFile().getName();
                 var ext = fileName.substring(fileName.lastIndexOf("."));
@@ -417,7 +417,7 @@ public class BatchModeEventListener implements ProcessingEventListener, ImageMat
         if (item.log().length() > 0 && header != null) {
             var namingStrategy = createNamingStrategy();
             var fileName = item.file().getName();
-            var logFileName = namingStrategy.render(sequenceNumber, null, "log", "notifications", fileName.substring(0, fileName.lastIndexOf("."))) + ".txt";
+            var logFileName = namingStrategy.render(sequenceNumber, null, "log", "notifications", fileName.substring(0, fileName.lastIndexOf(".")), null) + ".txt";
             try {
                 var logFilePath = outputDirectory.toPath().resolve(logFileName);
                 Files.writeString(logFilePath, item.log().toString(), Charset.defaultCharset());